+++ Procfile
@@ -1,1 +1,1 @@
-web: java -jar backend/target/equipment_tanks-app-1.0.4-SNAPSHOT.jar
+web: java -jar backend/target/tanks-and-equipments-app-1.0.4-SNAPSHOT.jar
