+++ backend/src/main/java/com/glassait/tanks_and_equipments/GlassaitApplication.java
@@ -1,13 +1,8 @@
-package com.glassait.equipment_tanks;
+package com.glassait.tanks_and_equipments;
 
-import com.glassait.equipment_tanks.config.DataSourceConfigProperties;
-import com.glassait.equipment_tanks.config.GlassaitConfigProperties;
-import com.glassait.equipment_tanks.config.SecurityConfigProperties;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 /**
  * Main class of the Glassait application.
@@ -15,8 +10,6 @@ import org.springframework.scheduling.annotation.EnableScheduling;
  * @author Glassait
  */
 @SpringBootApplication
-@EnableConfigurationProperties({ DataSourceConfigProperties.class, SecurityConfigProperties.class, GlassaitConfigProperties.class })
-@EnableScheduling
 @Slf4j
 public class GlassaitApplication {
     /**
