+++ backend/src/main/resources/application.properties
@@ -5,6 +5,6 @@ spring.datasource.hikari.maximum-pool-size=5
 #Set logging level for specific packages/classes
 logging.level.org.springframework=INFO
 logging.level.org.hibernate=WARN
-logging.level.com.glassait.equipment_tanks=DEBUG
+logging.level.com.glassait.tanks-and-equipments=DEBUG
 # Heroku
 server.port=${PORT:5000}
