+++ pom.xml
@@ -11,13 +11,20 @@
     </parent>
 
     <groupId>com.glassait</groupId>
-    <artifactId>equipment_tanks-parent</artifactId>
+    <artifactId>tanks-and-equipments-parent</artifactId>
     <version>${revision}</version>
     <description>Backend for equipmentTanks web site</description>
 
     <properties>
-        <java.version>20</java.version>
         <revision>2.0.0-SNAPSHOT</revision>
+
+        <java.version>20</java.version>
+
+        <version.thingsboard>3.0.4</version.thingsboard>
+
+        <version.spring-boot>3.2.2</version.spring-boot>
+        <version.lombok>1.18.30</version.lombok>
+        <version.mapstruct>1.5.5.Final</version.mapstruct>
     </properties>
 
     <packaging>pom</packaging>
